+++ ouds_core/lib/components/control/ouds_control_item.dart
@@ -178,7 +178,7 @@ class OudsControlItemState extends State<OudsControlItem> {
                   borderRadius: BorderRadius.circular(borderTokens.radiusNone),
                 ),
                 constraints: BoxConstraints(
-                  minHeight: controlItemTokens.sizeMinHeight,
+                  minHeight: controlItemTokens.sizeMinHeightDefault,
                   minWidth: controlItemTokens.sizeMinWidth,
                   maxWidth: widget.constrainedMaxWidth
                       ? controlItemTokens.sizeMaxWidth
@@ -263,7 +263,7 @@ class OudsControlItemState extends State<OudsControlItem> {
             Padding(
               padding: EdgeInsetsDirectional.only(
                 start: controlItemTokens.spacePaddingInline,
-                top: controlItemTokens.spacePaddingBlockTopErrorText,
+                top: controlItemTokens.spacePaddingBlockTopHelperText,
                 end: controlItemTokens.spacePaddingInline,
               ),
               child: Text(
@@ -470,6 +470,9 @@ class OudsControlItemState extends State<OudsControlItem> {
           maxHeight: OudsTheme.of(
             context,
           ).componentsTokens(context).controlItem.sizeMaxHeightAssetsContainer,
+          minHeight: OudsTheme.of(
+            context,
+          ).componentsTokens(context).controlItem.sizeIcon,
         ),
         alignment: Alignment.center,
         child: SizedBox(